feat: Implement in memory EvalSetsManager
This version of the EvalSetsManager is intended to support behavior where the agent developer wants to bring in their own eval sets, which is usually the case with `adk eval` cli. Once their eval sets are uploaded into this version of the eval sets manager, the EvalSetManager could be handed over to the Eval system for running evals.

PiperOrigin-RevId: 782197052